requirements.txt
Django==1.11.2
django-allauth==0.29.0
django-ckeditor==5.2.2
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
django-rest-swagger==2.1.0
django-widget-tweaks==1.4.1
sparkpost==1.3.5
psycopg2==2.7
raven==6.2.1
django-webpack-loader==0.4.1
